requirements_bot.txt
aiogram==3.7.0
telethon==1.42.0
SQLAlchemy==2.0.31
asyncpg==0.29.0
python-dotenv==1.0.1
qrcode[pil]==7.4.2
